Woods - little update
Changes in the woods level. Added textures assets for trees, planks etc. Create a prefab platform for the woods level and its moving variation and create a variation for the goblin prefab (scale down).
Changes in the woods scene:
1. Extending terrain & water
2. Added some platforms and enemies
3. Added first moving platform
It is only the preview of future update of this level.